# Optimize `CowData`'s `remove_at` and `insert` functions to avoid repeated copy assignments for non-trivial types.