@RandomShaper 2024-12-16
Looks like this is neglecting cases where `T` is not trivially copyable.