@Ivorforce 2024-12-16
> Looks like this is neglecting cases where `T` is not trivially copyable.

Are you sure a check for that is required? No copies are taking place, implicit or explicit - it's all moves, construction and destruction.