@Ivorforce 2025-10-01
Closing this because the code has been optimized in a better way already, with #105928.